requirements.txt
SQLAlchemy
psycopg2-binary
pandas
scikit-learn
pulp
python-dotenv
openpyxl